refactored index.d.ts typescript definitions into various files for faster auditing and management
