+++ lib/terminal/utilities/readStorage.ts
@@ -11,7 +11,7 @@ const readStorage = function terminal_utilities_readStorage(callback:(storage:st
             return;
         }
         let length:number = fileList.length;
-        const flag:storageFlag = {},
+        const flag:flagList = {},
             storage:storageItems = {
                 device: {},
                 settings: {
