+++ lib/typescript/readme.md
@@ -1,0 +1,14 @@
+# lib/typescript Code Files
+The various TypeScript type and interface definitions
+
+## Files
+<!-- Do not edit below this line.  Contents dynamically populated. -->
+
+* **[browser.d.ts](browser.d.ts)**                 - TypeScript interfaces used by browser specific libraries.
+* **[browser_modules.d.ts](browser_modules.d.ts)** - TypeScript interfaces that define master library modules used in the browser.
+* **[environment.d.ts](environment.d.ts)**         - TypeScript interfaces that define environmental objects.
+* **[fileService.d.ts](fileService.d.ts)**         - TypeScript interfaces used by the file services.
+* **[global.d.ts](global.d.ts)**                   - TypeScript interfaces used in many unrelated areas of the application.
+* **[terminal.d.ts](terminal.d.ts)**               - TypeScript interfaces used by terminal specific libraries.
+* **[test.d.ts](test.d.ts)**                       - TypeScript interfaces used test automation.
+* **[types.d.ts](types.d.ts)**                     - TypeScript static types.
